requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
cffi==1.12.3
decorator==4.4.0
Django==2.2.5
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.9.4
dj_database_url==0.5.0
ecdsa==0.13.2
future==0.17.1
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jose==1.0.0
Markdown==3.1.1
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2
psycopg2-binary
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.8.2
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-jose==2.0.2

requests==2.13.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
